int auf double korrigiert

+++ Taschenrechner/src/TaschenRechner.java
@@ -5,8 +5,8 @@ public class TaschenRechner
 
     public static void main(String[] args)
     {
-        int num1;
-        int num2;
+        double num1;
+        double num2;
         String operation;
